split zmq runner into subpackage
